Add markdown and CSV output
