+++ test/shared/test_compliance_data.py
@@ -4,6 +4,11 @@ import json
 from azure_guardrails.shared import utils
 from azure_guardrails.scrapers.compliance_data import PolicyComplianceData, ComplianceCoverage, PolicyDefinitionMetadata, \
     BenchmarkEntry
+import unittest
+from azure_guardrails.shared import utils
+from azure_guardrails.scrapers.compliance_data import ComplianceCoverage
+from azure_guardrails.guardrails.services import Services, Service
+from azure_guardrails.shared.config import get_default_config, get_config_from_file
 
 
 class PolicyComplianceDataTestCase(unittest.TestCase):
@@ -44,3 +49,35 @@ class ComplianceCoverageTestCase(unittest.TestCase):
         markdown_table = compliance_coverage.markdown_table()
         # print(markdown_table)
         # print(tabulate(results, headers=headers, tablefmt="github"))
+
+
+class OtherComplianceCoverageTestCase(unittest.TestCase):
+    def setUp(self):
+        service = "all"
+        config = get_default_config(exclude_services=None)
+        with_parameters = False
+        if service == "all":
+            self.services = Services(config=config)
+            self.policy_names = self.services.get_display_names(with_parameters=with_parameters)
+        else:
+            self.services = Service(service_name=service, config=config)
+            self.policy_names = self.services.get_display_names(with_parameters=with_parameters)
+        self.compliance_coverage = ComplianceCoverage(display_names=self.policy_names)
+
+    def test_markdown_table(self):
+        markdown_table = self.compliance_coverage.markdown_table()
+        summary = self.compliance_coverage.table_summary()
+        print(markdown_table)
+
+    # def test_csv_table(self):
+    #     path = os.path.join(
+    #         os.path.pardir,
+    #         os.path.pardir,
+    #         "test.csv"
+    #     )
+    #     if os.path.exists(path):
+    #         print("Removing the previous file")
+    #         os.remove(path)
+    #     self.compliance_coverage.csv_table(path, verbosity=0)
+    #     print(f"CSV updated! Path: {path}")
+
